Start of MP5, refactoring of codes into modular style and adding OOP Features
